control_service/requirements.txt
sqlalchemy
schema
fastapi
databases[postgresql]
aiosqlite
passlib[bcrypt]
python-multipart
pyjwt
aiodns
